refactor: change tests naming
