+++ react-basic/src/components/CardList/CardList.test.tsx
@@ -1,26 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
+import { generateComponent, testCardsData } from '../../test/mock/cardListMock';
 
-import { BrowserRouter } from 'react-router-dom';
-import { cardListTestProps, testCardsData } from '../../test/mockDataForTest';
-import { CardsContext } from '../../context/cards-context';
-import CardList from './CardList';
-import { SearchContext } from '../../context/search-context';
-import { ICard } from '../../types/card-type';
-
-describe('Renders CardList correctly', () => {
-  function generateComponent(cards: ICard[] | []): JSX.Element {
-    return (
-      <BrowserRouter>
-        <SearchContext.Provider value={{ search: '', changeSearch: () => {} }}>
-          <CardsContext.Provider value={{ cards, setCards: () => {} }}>
-            <CardList {...cardListTestProps} />
-          </CardsContext.Provider>
-        </SearchContext.Provider>
-      </BrowserRouter>
-    );
-  }
-
+describe('Tests for the Card List component', () => {
   it('Renders the specified number of cards', async () => {
     render(generateComponent(testCardsData));
