+++ react-basic/src/pages/ErrorPage/ErrorPage.test.tsx
@@ -3,18 +3,12 @@ import { render, screen } from '@testing-library/react';
 import ErrorPage from './ErrorPage';
 import { createMemoryRouter, RouterProvider } from 'react-router-dom';
 
-describe('Renders error-page correctly', async () => {
+describe('Tests for the 404 Page component', async () => {
   const router = createMemoryRouter([{ path: '/', element: <ErrorPage /> }], {
     initialEntries: ['/'],
   });
 
-  it('Should render the Error page correctly', async () => {
-    render(<RouterProvider router={router} />);
-    const heading = await screen.queryByText('Oops!');
-    expect(heading).not.toBeNull();
-  });
-
-  it('Should render 404 Page correctly', async () => {
+  it('Show 404 page when navigating to an invalid route', async () => {
     render(<RouterProvider router={router} />);
     const profileList = screen.getByTestId('errorMessage');
     expect(profileList).toBeInTheDocument();
