+++ react-basic/src/test/mock/cardMock.ts
@@ -1,0 +1,10 @@
+import { CardProps } from '../../components/Card/Card';
+import { PAGE, PAGE_SIZE } from '../../utils/constants';
+
+export const cardTestProps: CardProps = {
+  id: '1',
+  image: 'https://images.pokemontcg.io/hgss4/1.png',
+  name: 'Aggron',
+  currentPage: PAGE,
+  currentPageSize: PAGE_SIZE,
+};
